issue-145 FeedItem -> Feed로 이름 변경, 기존의 값은 임시로 변경
