+++ app/src/main/java/com/boostcamp/dreampicker/data/model/UserDetail.java
@@ -14,7 +14,7 @@ public class UserDetail {
 
     private List<User> followerList; // 해당 유저의 팔로워 리스트
     private List<User> followingList; // 해당 유저의 팔로잉 리스트
-    private List<Feed> feedList; // 해당 유저의 피드 리스트
+    private List<FeedPrevious> feedList; // 해당 유저의 피드 리스트
 
     public UserDetail() { }
 
@@ -116,11 +116,11 @@ public class UserDetail {
         this.followingList = followingList;
     }
 
-    public List<Feed> getFeedList() {
+    public List<FeedPrevious> getFeedList() {
         return feedList;
     }
 
-    public void setFeedList(List<Feed> feedList) {
+    public void setFeedList(List<FeedPrevious> feedList) {
         this.feedList = feedList;
     }
 }
