+++ app/src/main/java/com/boostcamp/dreampicker/data/paging/datasource/ProfileFeedDataSource.java
@@ -2,7 +2,7 @@ package com.boostcamp.dreampicker.data.paging.datasource;
 
 import android.annotation.SuppressLint;
 
-import com.boostcamp.dreampicker.data.model.Feed;
+import com.boostcamp.dreampicker.data.model.FeedPrevious;
 import com.boostcamp.dreampicker.data.source.remote.firebase.response.PagedListResponse;
 import com.boostcamp.dreampicker.data.source.repository.FeedRepository;
 
@@ -12,7 +12,7 @@ import androidx.paging.DataSource;
 import androidx.paging.PageKeyedDataSource;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 
-public class ProfileFeedDataSource extends PageKeyedDataSource<Integer, Feed> {
+public class ProfileFeedDataSource extends PageKeyedDataSource<Integer, FeedPrevious> {
 
     @NonNull
     private FeedRepository repository;
@@ -30,9 +30,9 @@ public class ProfileFeedDataSource extends PageKeyedDataSource<Integer, Feed> {
     @SuppressLint("CheckResult")
     @Override
     public void loadInitial(@NonNull LoadInitialParams<Integer> params,
-                            @NonNull LoadInitialCallback<Integer, Feed> callback) {
+                            @NonNull LoadInitialCallback<Integer, FeedPrevious> callback) {
 
-        PagedListResponse<Feed> response = repository.addProfileFeedList(userId, 1, params.requestedLoadSize)
+        PagedListResponse<FeedPrevious> response = repository.addProfileFeedList(userId, 1, params.requestedLoadSize)
                 .observeOn(AndroidSchedulers.mainThread())
                 .doOnError(error -> {})
                 .blockingGet();
@@ -49,17 +49,17 @@ public class ProfileFeedDataSource extends PageKeyedDataSource<Integer, Feed> {
     @SuppressLint("CheckResult")
     @Override
     public void loadBefore(@NonNull LoadParams<Integer> params,
-                           @NonNull LoadCallback<Integer, Feed> callback) {
+                           @NonNull LoadCallback<Integer, FeedPrevious> callback) {
         // ignore
     }
 
     @SuppressLint("CheckResult")
     @Override
     public void loadAfter(@NonNull LoadParams<Integer> params,
-                          @NonNull LoadCallback<Integer, Feed> callback) {
+                          @NonNull LoadCallback<Integer, FeedPrevious> callback) {
 
         if(!isPageEnd){
-            PagedListResponse<Feed> response = repository.addProfileFeedList(userId, params.key, params.requestedLoadSize)
+            PagedListResponse<FeedPrevious> response = repository.addProfileFeedList(userId, params.key, params.requestedLoadSize)
                     .observeOn(AndroidSchedulers.mainThread())
                     .doOnError(error -> {})
                     .blockingGet();
@@ -74,7 +74,7 @@ public class ProfileFeedDataSource extends PageKeyedDataSource<Integer, Feed> {
     }
 
 
-    public static class Factory extends DataSource.Factory<Integer, Feed> {
+    public static class Factory extends DataSource.Factory<Integer, FeedPrevious> {
 
         @NonNull
         private final FeedRepository repository;
@@ -93,7 +93,7 @@ public class ProfileFeedDataSource extends PageKeyedDataSource<Integer, Feed> {
 
         @NonNull
         @Override
-        public DataSource<Integer, Feed> create() {
+        public DataSource<Integer, FeedPrevious> create() {
             ProfileFeedDataSource source = new ProfileFeedDataSource(repository, userId);
             liveData.postValue(source);
             return source;
