+++ app/src/main/java/com/boostcamp/dreampicker/data/source/FeedDataSource.java
@@ -1,10 +1,8 @@
 package com.boostcamp.dreampicker.data.source;
 
-import com.boostcamp.dreampicker.data.model.Feed;
+import com.boostcamp.dreampicker.data.model.FeedPrevious;
 import com.boostcamp.dreampicker.data.source.remote.firebase.response.PagedListResponse;
 import com.boostcamp.dreampicker.utils.Constant;
-
-import java.util.List;
 
 import androidx.annotation.NonNull;
 import io.reactivex.Completable;
@@ -14,7 +12,7 @@ public interface FeedDataSource {
 
      // [메인 피드] 내가 투표 안하고, 마감 안된 투표 최신순으로 페이징
     @NonNull
-    Single<PagedListResponse<Feed>> addMainFeedList(int start, int display);
+    Single<PagedListResponse<FeedPrevious>> addMainFeedList(int start, int display);
 
     // [피드공통] 투표하기 / 투표변경 / 투표취소
     @NonNull
@@ -22,15 +20,15 @@ public interface FeedDataSource {
 
     // [검색] 검색 결과 피드 리스트 페이징
     @NonNull
-    Single<PagedListResponse<Feed>> addSearchFeedList(@NonNull String searchKey,
-                                                      int start,
-                                                      int display);
+    Single<PagedListResponse<FeedPrevious>> addSearchFeedList(@NonNull String searchKey,
+                                                              int start,
+                                                              int display);
 
     //[프로필] 해당 유저가 업로드한 피드 리스트 페이징
     @NonNull
-    Single<PagedListResponse<Feed>> addProfileFeedList(@NonNull String userId,
-                                                       int start,
-                                                       int display);
+    Single<PagedListResponse<FeedPrevious>> addProfileFeedList(@NonNull String userId,
+                                                               int start,
+                                                               int display);
 
     // [피드공통] 해당 투표 종료/재시작 토글
     @NonNull
@@ -39,5 +37,5 @@ public interface FeedDataSource {
 
     // [업로드]
     @NonNull
-    Completable upLoadFeed(@NonNull Feed feed);
+    Completable upLoadFeed(@NonNull FeedPrevious feed);
 }
