+++ app/src/main/java/com/boostcamp/dreampicker/presentation/profile/ProfileFeedAdapter.java
@@ -5,7 +5,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.boostcamp.dreampicker.R;
-import com.boostcamp.dreampicker.data.model.Feed;
+import com.boostcamp.dreampicker.data.model.FeedPrevious;
 import com.boostcamp.dreampicker.databinding.ItemProfileFeedBinding;
 import com.boostcamp.dreampicker.presentation.BasePagedListAdapter;
 import com.boostcamp.dreampicker.presentation.listener.OnItemClickListener;
@@ -16,9 +16,9 @@ import androidx.databinding.DataBindingUtil;
 import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class ProfileFeedAdapter extends BasePagedListAdapter<Feed, ProfileFeedAdapter.ViewHolder> {
+public class ProfileFeedAdapter extends BasePagedListAdapter<FeedPrevious, ProfileFeedAdapter.ViewHolder> {
 
-    public ProfileFeedAdapter(@Nullable OnItemClickListener<Feed> onItemClickListener) {
+    public ProfileFeedAdapter(@Nullable OnItemClickListener<FeedPrevious> onItemClickListener) {
         super(DIFF_CALLBACK, onItemClickListener);
     }
 
@@ -32,7 +32,7 @@ public class ProfileFeedAdapter extends BasePagedListAdapter<Feed, ProfileFeedAd
     @Override
     public void onBindView(@NonNull ViewHolder holder, int position) {
 
-        final Feed item = getItem(position);
+        final FeedPrevious item = getItem(position);
 
         holder.bindTo(item);
     }
@@ -46,20 +46,20 @@ public class ProfileFeedAdapter extends BasePagedListAdapter<Feed, ProfileFeedAd
             binding = DataBindingUtil.bind(itemView);
         }
 
-        void bindTo(Feed feed) {
+        void bindTo(FeedPrevious feed) {
             binding.setItem(feed);
         }
     }
 
-    private static final DiffUtil.ItemCallback<Feed> DIFF_CALLBACK =
-            new DiffUtil.ItemCallback<Feed>() {
+    private static final DiffUtil.ItemCallback<FeedPrevious> DIFF_CALLBACK =
+            new DiffUtil.ItemCallback<FeedPrevious>() {
                 @Override
-                public boolean areItemsTheSame(@NonNull Feed oldItem, @NonNull Feed newItem) {
+                public boolean areItemsTheSame(@NonNull FeedPrevious oldItem, @NonNull FeedPrevious newItem) {
                     return oldItem.getId().equals(newItem.getId());
                 }
 
                 @Override
-                public boolean areContentsTheSame(@NonNull Feed oldItem, @NonNull Feed newItem) {
+                public boolean areContentsTheSame(@NonNull FeedPrevious oldItem, @NonNull FeedPrevious newItem) {
                     return oldItem.equals(newItem);
                 }
             };
