+++ app/src/main/res/layout/item_feed.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="feed"
-            type="com.boostcamp.dreampicker.data.model.Feed"/>
+            type="com.boostcamp.dreampicker.data.model.FeedPrevious"/>
     </data>
 
     <LinearLayout
